添加到点不旋转功能、添加重新打开app加载上一次位置点功能

+++ TobotApp/src/main/java/com/tobot/map/module/main/map/AddPointViewDialog.java
@@ -174,14 +174,16 @@ public class AddPointViewDialog extends BaseAnimDialog implements View.OnClickLi
     }
 
     @Override
-    public void onConfirm() {
-        // 删除确认的操作
-        if (mLocationBean != null) {
-            String number = mLocationBean.getLocationNumber();
-            MyDBSource.getInstance(getActivity()).deleteLocation(number);
-            showLocationData(MyDBSource.getInstance(getActivity()).queryLocationList());
-            if (mOnPointListener != null) {
-                mOnPointListener.onDeleteLocationLabel(number);
+    public void onConfirm(boolean isConfirm) {
+        if (isConfirm) {
+            // 删除确认的操作
+            if (mLocationBean != null) {
+                String number = mLocationBean.getLocationNumber();
+                MyDBSource.getInstance(getActivity()).deleteLocation(number);
+                showLocationData(MyDBSource.getInstance(getActivity()).queryLocationList());
+                if (mOnPointListener != null) {
+                    mOnPointListener.onDeleteLocationLabel(number);
+                }
             }
         }
     }
